Imported Upstream version 1.10.4

+++ fish/edit.c
@@ -1,5 +1,5 @@
 /* guestfish - the filesystem interactive shell
- * Copyright (C) 2009 Red Hat Inc.
+ * Copyright (C) 2009-2011 Red Hat Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,62 +24,21 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <inttypes.h>
+#include <sys/types.h>
+#include <sys/stat.h>
 
 #include "fish.h"
 
 /* guestfish edit command, suggested by Ján Ondrej, implemented by RWMJ */
 
-static char *
-load_file (const char *filename, size_t *len_r)
-{
-  int fd, r, start;
-  char *content = NULL, *p;
-  char buf[65536];
-
-  *len_r = 0;
-
-  fd = open (filename, O_RDONLY);
-  if (fd == -1) {
-    perror (filename);
-    return NULL;
-  }
-
-  while ((r = read (fd, buf, sizeof buf)) > 0) {
-    start = *len_r;
-    *len_r += r;
-    p = realloc (content, *len_r + 1);
-    if (p == NULL) {
-      perror ("realloc");
-      free (content);
-      return NULL;
-    }
-    content = p;
-    memcpy (content + start, buf, r);
-    content[start+r] = '\0';
-  }
-
-  if (r == -1) {
-    perror (filename);
-    free (content);
-    return NULL;
-  }
-
-  if (close (fd) == -1) {
-    perror (filename);
-    free (content);
-    return NULL;
-  }
-
-  return content;
-}
-
 int
-do_edit (const char *cmd, int argc, char *argv[])
+run_edit (const char *cmd, size_t argc, char *argv[])
 {
-  char filename[] = "/tmp/guestfishXXXXXX";
+  TMP_TEMPLATE_ON_STACK (filename);
   char buf[256];
   const char *editor;
-  char *content, *content_new;
+  char *remotefilename;
+  struct stat oldstat, newstat;
   int r, fd;
 
   if (argc != 1) {
@@ -98,30 +57,40 @@ do_edit (const char *cmd, int argc, char *argv[])
       editor = "vi"; /* could be cruel here and choose ed(1) */
   }
 
+  /* Handle 'win:...' prefix. */
+  remotefilename = win_prefix (argv[0]);
+  if (remotefilename == NULL)
+    return -1;
+
   /* Download the file and write it to a temporary. */
   fd = mkstemp (filename);
   if (fd == -1) {
     perror ("mkstemp");
+    free (remotefilename);
     return -1;
   }
 
-  if ((content = guestfs_cat (g, argv[0])) == NULL) {
-    close (fd);
-    unlink (filename);
-    return -1;
-  }
+  snprintf (buf, sizeof buf, "/dev/fd/%d", fd);
 
-  if (xwrite (fd, content, strlen (content)) == -1) {
+  if (guestfs_download (g, remotefilename, buf) == -1) {
     close (fd);
     unlink (filename);
-    free (content);
+    free (remotefilename);
     return -1;
   }
 
   if (close (fd) == -1) {
     perror (filename);
     unlink (filename);
-    free (content);
+    free (remotefilename);
+    return -1;
+  }
+
+  /* Get the old stat. */
+  if (stat (filename, &oldstat) == -1) {
+    perror (filename);
+    unlink (filename);
+    free (remotefilename);
     return -1;
   }
 
@@ -133,36 +102,34 @@ do_edit (const char *cmd, int argc, char *argv[])
   if (r != 0) {
     perror (buf);
     unlink (filename);
-    free (content);
+    free (remotefilename);
     return -1;
   }
 
-  /* Reload it. */
-  size_t size;
-  content_new = load_file (filename, &size);
-  if (content_new == NULL) {
+  /* Get the new stat. */
+  if (stat (filename, &newstat) == -1) {
+    perror (filename);
     unlink (filename);
-    free (content);
+    free (remotefilename);
     return -1;
   }
-
-  unlink (filename);
 
   /* Changed? */
-  if (strlen (content) == size && STREQLEN (content, content_new, size)) {
-    free (content);
-    free (content_new);
+  if (oldstat.st_ctime == newstat.st_ctime &&
+      oldstat.st_size == newstat.st_size) {
+    unlink (filename);
+    free (remotefilename);
     return 0;
   }
 
   /* Write new content. */
-  if (guestfs_write_file (g, argv[0], content_new, size) == -1) {
-    free (content);
-    free (content_new);
+  if (guestfs_upload (g, filename, remotefilename) == -1) {
+    unlink (filename);
+    free (remotefilename);
     return -1;
   }
 
-  free (content);
-  free (content_new);
+  unlink (filename);
+  free (remotefilename);
   return 0;
 }
